feat: Rename folder and improve style
