+++ source/components/groupe/Button.tsx
@@ -12,7 +12,7 @@ export default function Button({
 	return (
 		<button
 			onClick={onClick}
-			className={`inline-flex items-center px-8 py-4 border border-transparent text-sm font-medium rounded-sm shadow-sm text-white bg-primary hover:opacity-0.7 focus:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-primary-600 ${className}`}
+			className={`inline-flex items-center px-8 py-4 border border-transparent text-sm font-medium rounded-sm shadow-sm text-white bg-primary hover:opacity-80 focus:text-white focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-primary-600 transition-opacity transition-colors ${className}`}
 			{...props}
 		>
 			{children}
