+++ source/pages/creer-groupe/InformationsGroupe.tsx
