+++ source/pages/creer-groupe/components/CopyInput.tsx
