+++ source/pages/creer-groupe/components/StepperIndicator.tsx
